requirements.txt
alembic==1.13.1
annotated-types==0.7.0
blinker==1.7.0
blis==0.7.11
catalogue==2.0.10
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.5
cymem==2.0.8
Flask==3.0.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
gunicorn==22.0.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.4.2
langcodes==3.4.0
language_data==1.2.0
Mako==1.3.0
marisa-trie==1.1.1
MarkupSafe==2.1.3
murmurhash==1.0.10
mysqlclient==2.2.1
nltk==3.8.1
numpy==1.26.4
packaging==24.0
preshed==3.0.9
pt-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/pt_core_news_sm-3.7.0/pt_core_news_sm-3.7.0-py3-none-any.whl#sha256=a3a8eed42c600af17c70ed3c8720ffe2879b5353055e66fe0c998306329052fd
pydantic==2.7.2
pydantic_core==2.18.3
python-dotenv==1.0.0
regex==2024.5.15
requests==2.31.0
setuptools==70.0.0
smart-open==6.4.0
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.25
srsly==2.4.8
thinc==8.2.3
tqdm==4.66.4
typer==0.9.4
typing_extensions==4.9.0
urllib3==2.1.0
wasabi==1.1.3
weasel==0.3.4
Werkzeug==3.0.1
wheel==0.43.0
